refactor: normalize kebab-case assets and drop unused ttf fonts
