+++ src/assets/fonts/inter/inter-bold.woff2
