+++ src/config/Fonts/inter.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: 'Inter';
-  src: url('../../assets/fonts/inter/Inter-Regular.woff2') format('woff2');
+  src: url('../../assets/fonts/inter/inter-regular.woff2') format('woff2');
   font-weight: 400;
   font-style: normal;
   font-display: swap;
@@ -8,7 +8,7 @@
 
 @font-face {
   font-family: 'Inter';
-  src: url('../../assets/fonts/inter/Inter-Medium.woff2') format('woff2');
+  src: url('../../assets/fonts/inter/inter-medium.woff2') format('woff2');
   font-weight: 500;
   font-style: normal;
   font-display: swap;
@@ -16,7 +16,7 @@
 
 @font-face {
   font-family: 'Inter';
-  src: url('../../assets/fonts/inter/Inter-Bold.woff2') format('woff2');
+  src: url('../../assets/fonts/inter/inter-bold.woff2') format('woff2');
   font-weight: 700;
   font-style: normal;
   font-display: swap;
