+++ src/styles/fonts.css
@@ -2,38 +2,38 @@
   font-family: 'Golos';
   font-display: swap;
   font-weight: 400;
-  src: url('../assets/fonts/golos/Golos-Text_Regular.woff2') format('woff2');
+  src: url('../assets/fonts/golos/golos-text-regular.woff2') format('woff2');
 }
 
 @font-face {
   font-family: 'Golos';
   font-display: swap;
   font-weight: 500;
-  src: url('../assets/fonts/golos/Golos-Text_Medium.woff2') format('woff2');
+  src: url('../assets/fonts/golos/golos-text-medium.woff2') format('woff2');
 }
 @font-face {
   font-family: 'Golos';
   font-display: swap;
   font-weight: 700;
-  src: url('../assets/fonts/golos/Golos-Text_Bold.woff2') format('woff2');
+  src: url('../assets/fonts/golos/golos-text-bold.woff2') format('woff2');
 }
 @font-face {
   font-family: 'Golos';
   font-display: swap;
   font-weight: 900;
-  src: url('../assets/fonts/golos/Golos-Text_Black.woff2') format('woff2');
+  src: url('../assets/fonts/golos/golos-text-black.woff2') format('woff2');
 }
 
 @font-face {
   font-family: 'Inter';
   font-display: swap;
   font-weight: 400;
-  src: url('../assets/fonts/inter/Inter-Regular.woff2') format('woff2');
+  src: url('../assets/fonts/inter/inter-regular.woff2') format('woff2');
 }
 
 @font-face {
   font-family: 'Inter';
   font-display: swap;
   font-weight: 500;
-  src: url('../assets/fonts/inter/Inter-Medium.woff2') format('woff2');
+  src: url('../assets/fonts/inter/inter-medium.woff2') format('woff2');
 }
