+++ tests/visual/get-registration-notification-snapshot-name.ts
