+++ tests/visual/visual-comparison.registration-notification.spec.ts-snapshots/uk-error-large-mobile-2-chromium-linux.png
