+++ tests/visual/visual-comparison.registration-notification.spec.ts-snapshots/uk-error-large-mobile-firefox-linux.png
