+++ tests/visual/visual-comparison.registration-notification.spec.ts-snapshots/uk-error-large-mobile-webkit-linux.png
