+++ tests/visual/visual-comparison.registration-notification.spec.ts-snapshots/uk-success-large-mobile-2-firefox-linux.png
