+++ tests/visual/visual-comparison.registration-notification.spec.ts-snapshots/uk-success-large-mobile-chromium-linux.png
