+++ tests/visual/visualComparison.authentication.spec.ts-snapshots/uk-large-mobile-2-webkit-linux.png
