+++ tests/visual/visualComparison.authentication.spec.ts-snapshots/uk-large-mobile-chromium-linux.png
