+++ tests/visual/visualComparison.authentication.spec.ts-snapshots/uk-large-mobile-firefox-linux.png
