+++ tests/visual/visualComparison.authentication.spec.ts-snapshots/uk-large-mobile-webkit-linux.png
